requirements.txt
Flask
Flask-Cors
pytz
langchain
pinecone-client
groq
openai